requirements.txt
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.25.3, < 2.1.0
pydantic >= 1.10.5, < 2
aenum >= 3.1.11
requests
pyyaml
python-dotenv # Only used for the examples, not required